+++ xrpl_types/src/serialize/mod.rs
@@ -1,7 +1,5 @@
-mod field_id;
 mod serialize;
 mod serializer;
 
-pub use field_id::*;
 pub use serialize::*;
 pub use serializer::*;
