+++ xrpl_types/src/types/mod.rs
@@ -1,9 +1,11 @@
 mod amount;
 mod currency_code;
+mod ledger_timestamp;
 mod primitive;
 mod transaction;
 
 pub use amount::*;
 pub use currency_code::*;
+pub use ledger_timestamp::*;
 pub use primitive::*;
 pub use transaction::*;
