+++ xrpl_types/src/types/transaction/common.rs
@@ -1,43 +1,6 @@
 use crate::serialize::{Serialize, SerializeArray, Serializer};
 use crate::{AccountId, Amount, Blob, DropsAmount, Hash256, UInt32};
 
-#[repr(u16)]
-#[derive(Clone, Copy, PartialEq, Eq, Debug, serde::Serialize, serde::Deserialize)]
-pub enum TransactionType {
-    // Discriminant values can be found at https://github.com/XRPLF/xrpl.js/blob/main/packages/ripple-binary-codec/src/enums/definitions.json
-    Payment = 0,
-    EscrowCreate = 1,
-    EscrowFinish = 2,
-    AccountSet = 3,
-    EscrowCancel = 4,
-    SetRegularKey = 5,
-    NickNameSet = 6,
-    OfferCreate = 7,
-    OfferCancel = 8,
-    Contract = 9,
-    TicketCreate = 10,
-    TicketCancel = 11,
-    SignerListSet = 12,
-    PaymentChannelCreate = 13,
-    PaymentChannelFund = 14,
-    PaymentChannelClaim = 15,
-    CheckCreate = 16,
-    CheckCash = 17,
-    CheckCancel = 18,
-    DepositPreauth = 19,
-    TrustSet = 20,
-    AccountDelete = 21,
-    SetHook = 22,
-    NFTokenMint = 25,
-    NFTokenBurn = 26,
-    NFTokenCreateOffer = 27,
-    NFTokenCancelOffer = 28,
-    NFTokenAcceptOffer = 29,
-    EnableAmendment = 100,
-    SetFee = 101,
-    UNLModify = 102,
-}
-
 #[derive(Debug, Clone)]
 pub struct Memo {
     pub memo_type: Blob,
@@ -77,11 +40,6 @@ impl TransactionCommon {
             txn_signature: None,
         }
     }
-}
-
-pub trait Transaction {
-    fn common(&self) -> &TransactionCommon;
-    fn common_mut(&mut self) -> &mut TransactionCommon;
 }
 
 impl Serialize for TransactionCommon {
