+++ xrpl_types/src/types/transaction/mod.rs
@@ -1,15 +1,48 @@
-mod account_delete;
-mod account_set;
 mod common;
-mod offer_cancel;
-mod offer_create;
-mod payment;
-mod trust_set;
+mod variants;
 
-pub use account_delete::*;
-pub use account_set::*;
 pub use common::*;
-pub use offer_cancel::*;
-pub use offer_create::*;
-pub use payment::*;
-pub use trust_set::*;
+pub use variants::*;
+
+/// XRPL transaction
+pub trait Transaction {
+    fn common(&self) -> &TransactionCommon;
+    fn common_mut(&mut self) -> &mut TransactionCommon;
+}
+
+#[repr(u16)]
+#[derive(Clone, Copy, PartialEq, Eq, Debug, serde::Serialize, serde::Deserialize)]
+pub enum TransactionType {
+    // Discriminant values can be found at https://github.com/XRPLF/xrpl.js/blob/main/packages/ripple-binary-codec/src/enums/definitions.json
+    Payment = 0,
+    EscrowCreate = 1,
+    EscrowFinish = 2,
+    AccountSet = 3,
+    EscrowCancel = 4,
+    SetRegularKey = 5,
+    NickNameSet = 6,
+    OfferCreate = 7,
+    OfferCancel = 8,
+    Contract = 9,
+    TicketCreate = 10,
+    TicketCancel = 11,
+    SignerListSet = 12,
+    PaymentChannelCreate = 13,
+    PaymentChannelFund = 14,
+    PaymentChannelClaim = 15,
+    CheckCreate = 16,
+    CheckCash = 17,
+    CheckCancel = 18,
+    DepositPreauth = 19,
+    TrustSet = 20,
+    AccountDelete = 21,
+    SetHook = 22,
+    NFTokenMint = 25,
+    NFTokenBurn = 26,
+    NFTokenCreateOffer = 27,
+    NFTokenCancelOffer = 28,
+    NFTokenAcceptOffer = 29,
+    EnableAmendment = 100,
+    SetFee = 101,
+    UNLModify = 102,
+}
