+++ xrpl_types/src/types/transaction/variants/offer_create.rs
